requirements.txt
loguru>=0.7.2
requests>=2.31.0
pyyaml>=6.0.1
numpy>=1.26.1
pillow>=10.3.0
configparser>=5.0.2
langdetect>=1.0.9
iirosebot>=1.7.1
aiohttp>=3.8.6
websockets>=12.0
setuptools>=68.0.0